Move locales.js to lib/constants
