+++ lib/constants/locales.js
